Added files via upload
Attempt at creating the Delaunay Triangulation of a unit square with n-holes that are m-sided regular polygons.   As of right now the program is not properly masking the triangles in the holes and the triangles are distorted because only the four corner points on the square are included in the set of points.